requirements.txt
apache-airflow==2.6.2
pandas
s3fs